refactor: add fk condition

+++ src/data/services/ProjectPhase/DeleteProjectPhase.service.js
@@ -1,8 +1,12 @@
-import { ProjectPhaseRepository } from '../../database/repositories';
+import {
+  ProjectPhaseRepository,
+  ProductRepository,
+} from '../../database/repositories';
 
 export class DeleteProjectPhaseService {
   async execute({ id_project_phase }) {
     const repository = new ProjectPhaseRepository();
+    const productRepository = new ProductRepository();
 
     const verifyProjectPhaseExists = await repository.findProjectPhaseById({
       id_project_phase,
@@ -12,6 +16,17 @@ export class DeleteProjectPhaseService {
       return {
         error: `Não há nenhuma Fase de projeto registrado com este ID -> ${id_project_phase}.`,
       };
+
+    const verifyFk = await productRepository.verifyProjectPhase({
+      id_project_phase,
+    });
+
+    if (verifyFk.length > 0) {
+      return {
+        error:
+          'Não foi possível excluir a Fase de projeto pois existem Produtos associados.',
+      };
+    }
 
     await repository.deleteProjectPhase({
       id_project_phase,
